Title: 토끼와 경주, Time: 531ms, Memory: 23MB, Status: Passed - CodeTree

+++ 240613/토끼와 경주/rabit-and-race.java
@@ -148,7 +148,7 @@ public class Main {
 		pq1.add(rabbits[idx]);
 	}
 	
-	static int totalSum;
+	static long totalSum;
 	static void scoring(int idx) {
 		rabbits[idx].score -= (rabbits[idx].y + rabbits[idx].x);
 		totalSum += (rabbits[idx].y + rabbits[idx].x);
